+++ android/ui-common/src/main/java/com/tikal/tmdb/MovieListTile.kt
@@ -100,7 +100,7 @@ fun MovieListTile(
                 )
                 RatingBar(
                     modifier = Modifier.padding(top = 8.dp),
-                    value = movie.voteAverage / 2,
+                    value = (movie.voteAverage / 2).toFloat(),
                     config = RatingBarConfig()
                         .isIndicator(true)
                         .numStars(5)
